requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pymongo==4.6.0
google-cloud-pubsub>=2.20.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.107.0
httpx==0.25.1

# AI/ML Libraries
scikit-learn==1.3.2
numpy==1.26.0
scipy==1.11.4
joblib==1.3.2
protobuf>=6.31.1,<8.0.0
aioredis==2.0.1
protobuf>=4.25.8,<5.0.0
python-jose==3.3.0
passlib==1.7.4
emails==0.6
requests==2.31.0
celery==5.3.4
redis==5.0.1
motor==3.3.2
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
mypy==1.7.1
